add: generic info icon
